refactor(hexa): Change domain folder placement
