+++ backend-express/interface/controllers/auth.ts
@@ -1,7 +1,7 @@
 import { Request, Response, NextFunction } from 'express';
 import jwt from 'jsonwebtoken';
 import dotenv from "dotenv"
-import { SetEnvError, UnauthorizedError } from '../../../core/domain/errors/main';
+import { SetEnvError, UnauthorizedError } from '../../../domain/errors/main';
 import { CustomJwtPayload } from '../../express';
 import { userRepository } from './user';
