+++ backend-express/interface/controllers/post.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response } from "express";
 import { PrismaPostRepository } from "../../../infrastructure/repositories/prisma-post";
-import { FindDbError, UnauthorizedError } from "../../../core/domain/errors/main";
-import { Post } from "../../../core/domain/entities/Post";
+import { FindDbError, UnauthorizedError } from "../../../domain/errors/main";
+import { Post } from "../../../domain/entities/Post";
 import { UserJWT } from "../../express";
 
 const postRepository = new PrismaPostRepository()
