+++ backend-express/interface/controllers/user.ts
@@ -5,9 +5,9 @@ import dotenv from "dotenv"
 
 
 import {ReadAll, ReadByEmail, ReadById} from "../../../core/application/usecases/atomic/user"
-import {FindDbError, InvalidUrlError, SetEnvError, UnauthorizedError} from "../../../core/domain/errors/main"
+import {FindDbError, InvalidUrlError, SetEnvError, UnauthorizedError} from "../../../domain/errors/main"
 import { PrismaUserRepository } from "../../../infrastructure/repositories/prisma-user";
-import { User } from "../../../core/domain/entities/User";
+import { User } from "../../../domain/entities/User";
 export const userRepository = new PrismaUserRepository()
 
 dotenv.config()
