+++ core/application/usecases/atomic/user.ts
@@ -1,6 +1,6 @@
-import { FindDbError } from "../../../domain/errors/main";
+import { FindDbError } from "../../../../domain/errors/main";
 import { UserRepository } from "../../../application/repositories/user";
-import { User } from "../../../domain/entities/User";
+import { User } from "../../../../domain/entities/User";
 // import { User } from ".../../core/domain/entities/User";
 
 abstract class UseCaseBase {
