+++ core/test/post/create.test.ts
@@ -2,7 +2,7 @@
 
 import { PostRepository } from "application/repositories/post";
 import { CreatePost } from "application/usecases/atomic/post";
-import { Post } from "domain/entities/Post";
+import { Post } from "../../../domain/entities/Post";
