+++ core/test/user/create.test.ts
@@ -1,6 +1,6 @@
 import { UserRepository } from "application/repositories/user";
 import { CreateUser } from "application/usecases/atomic/user";
-import { User } from "domain/entities/User";
+import { User } from "../../../domain/entities/User";
 
 
 // Mock del repositorio
