+++ domain/entities/LikePost.ts
