+++ domain/entities/Post.ts
