+++ domain/errors/main.ts
