+++ infrastructure/repositories/prisma-likepost.ts
@@ -1,8 +1,9 @@
 import { PrismaClientConfig } from "../connectors/prisma-db";
 import { LikePostRepository } from "../../core/application/repositories/likepost";
+import { LikePost } from "../../domain/entities/LikePost";
 
 export class PrismaLikePostRepository extends PrismaClientConfig implements LikePostRepository {
-    async create(postId: number, userId: number): Promise<Object> {
+    async create(postId: number, userId: number): Promise<LikePost> {
         return await this.prisma.likePost.create({
             data: {
                 post: {
