+++ infrastructure/tsconfig.json
@@ -9,6 +9,6 @@
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true
   },
-  "include": ["./prisma", "./repositories", "./connectors"],
+  "include": ["./prisma", "./repositories", "./connectors", "../domain"],
   "exclude": ["node_modules"]
 }
